Split PcieAxiVersion class into its own file
