+++ python/axipcie/__init__.py
@@ -2,3 +2,4 @@
 
 from axipcie._AxiPcieCore import *
 from axipcie._AxiPipCore  import *
+from axipcie._PcieAxiVersion import *
